backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pandas
qdrant-client
dspy
google-generativeai
